requirements.txt
cairocffi==1.0.2
CairoSVG==2.3.0
cffi==1.12.2
cssselect2==0.2.1
cycler==0.10.0
defusedxml==0.5.0
kiwisolver==1.0.1
lxml==4.3.3
matplotlib==3.0.3
numpy==1.16.2
Pillow==6.0.0
pycparser==2.19
pyparsing==2.3.1
python-dateutil==2.8.0
reportlab==3.5.17
scikit-learn==0.20.3
scikit-optimize==0.5.2
scipy==1.2.1
six==1.12.0
svglib==0.9.0
svgwrite==1.2.1
tinycss2==1.0.2
webencodings==0.5.1
